// src/contexts/TaskContext/index.tsx
import { createContext, useContext, useState } from 'react';
import type { TaskStateModel } from '../../models/TaskStateModel';
const initialState: TaskStateModel = {
  tasks: [],
  secondsRemaining: 0,
  formattedSecondsRemaining: '00:00',
  activeTask: null,
  currentCycle: 0,
  config: {
    workTime: 25,
    shortBreakTime: 5,
    longBreakTime: 15,
  },
};

type TaskContextProps = {
  state: TaskStateModel; //dados globais
  setState: React.Dispatch<React.SetStateAction<TaskStateModel>>; //altera os dados globais
};

const initialContextValue = {
  state: initialState,
  setState: () => {},
};

export const TaskContext = createContext<TaskContextProps>(initialContextValue); //Criação do Context

type TaskContextProviderProps = {
  children: React.ReactNode;
};

//Criação do Provider | fornece o contexto para os componentes filhos
export function TaskContextProvider({ children }: TaskContextProviderProps) {
  const [state, setState] = useState(initialState);

  return (
    <TaskContext.Provider value={{ state, setState }}>
      {children}
    </TaskContext.Provider>
  );
}

//consome os dados do contexto
export function useTaskContext() {
  return useContext(TaskContext);
}
